refactor to use github pages as deployment target
