+++ .gitignore
@@ -51,8 +51,8 @@ ubuntu*.log
 # EEPROM image file (created by SITL)
 eeprom.bin
 
-# public file staging location 
-/public
+# output directory
+/output
 
 # NPM packages folder
 /node_modules
