--- index_hash.txt
@@ -1,1 +1,0 @@
-6206a3e6981c6564b16c4c4f2bc06fc8b75c837c1139bdd7c77ad3375eef987f
