--- indexer/check.js
@@ -1,6 +1,0 @@
-'use strict';
-
-const child = require('child_process').fork('indexer/indexer.js', ["nosave"]);
-child.on('exit', (code) => {
-    process.exitCode = code;
-});
