--- package.json
@@ -1,18 +1,0 @@
-{
-	"name" : "firmware-presets",
-	"productName" : "Betaflight Firmware Presets",
-	"description" : "Collection of CLI-based presets",
-	"scripts" : {
-		"index" : "node indexer/indexer.js",
-		"verify" : "node indexer/check.js"
-	},
-	"repository" : {
-		"type" : "git",
-		"url" : "github.com/betaflight/firmware-presets"
-	},
-	"author" : "The Betaflight open source project.",
-	"license" : "GPL-3.0",
-	"engines" : {
-		"node" : "10.x"
-	}
-}
